pyproject.toml
[project]
name = "RepulsBot"
version = "1"
description = "The official bot for the repuls.io discord server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asqlite>=2.0.0",
    "discord-py>=2.6.2",
    "dotenv>=0.9.9",
    "psutil>=7.1.3",
]